requirements.txt
bluepy
pycryptodomex